Move spells module
